refactor: use camelCase for filenames
